PythonOpenGL/requirements.txt
pyopengl
pyopengl-accelerate